requirements.txt
flask
flask-cors
flask-limiter
pymongo
python-dotenv
openai
gunicorn
